usb: qti_gadget: Fix the compilation warning on declaration after code
Currently, in the qti_usb_func_alloc function, there is a case
where 'qg' variable is defined after the code if CONFIG_USB_F_GSI
is enabled. This gives a compilation warning about mixed
declarations and code.

Fix this by declaring 'qg' inside the if block where it is used,
thereby limiting its scope to block.

Adding module license and descrption to the driver.

Change-Id: Icada05c50c1bb924ebeb362a2499bcd5fcf0b1ae
Signed-off-by: Shashikala Katthi <quic_skatthi@quicinc.com>