Serial: msm_geni_serial: Add RX wakeup support to UART console
Add the debug UART Rx pin as wakeup source for SA8155 target. When
we are in system suspend and to exit from the system suspend
earlier we were using the power key button. But with this patch
if we enter anything on UART console then it will act as wakeup
signal and system will resume. This console wakeup is used for
specific requirement to resume system when in complete suspend
state. To enable this feature we need to add the "qcom,console-
rx-wakeup" property in console UART dtsi node.

Change-Id: Ie480b50c0a09a449e11dfd9d1cb353677817fe7d
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>