Platform: msm: Read the property before the initalization
Reorder parameters parsing for ssr to work on linux guest.

Currently we are reading the "ssr-enable" property after the
geni_se_resource_init function and due to that when we are trying
to read the value it is NULL so to avoid this move the if check
before the function.

Change-Id: I5383850cac723d31540db022f1a64bebe8384a38
Signed-off-by: LADI RAM SAI <quic_lramsai@quicinc.com>