oplus_chg: Implement bq27541's time to full estimate and send to sysfs
This should fix the permanent 1hr, 24mins until full message on the
lockscreen.

Thankfully this was easy to implement, as the bq27541 has a
TimeToFull I2C command that should work. Datasheet here:
https://www.ti.com/lit/ds/symlink/bq27541.pdf

Thanks to Christian Oder for the sanity checks and fixups to this patch,
as well as verifying the behavior on retail hardware.

Co-authored-by: Christian Oder <myself5@carbonrom.org>
Signed-off-by: Jacob McSwain <jacob@mcswain.dev>
Change-Id: If935507e8e7085fd0bba0e9c2033888819b6ac29
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>