dmaengine: gpi: Reset Event and Tx,RX channels to start point
This change is to bring the GPII channels [Tx, RX] to the reset
point and also reset the Event channel.

When system enters into deep sleep and exits, this path to gpi
config need to be in similar state to the probe. So that we are
in same state post or during deep sleep exit path.

Change-Id: I80cd0ecf61cb004f4ec10dfde5dedc81aace81bb
Signed-off-by: Mukesh Kumar Savaliya <quic_msavaliy@quicinc.com>