serial: msm_geni_serial: Do not start Rx if port setup not done in GSI
Currently we are doing initialization of workqueue in port setup function.
port setup will happen when port open request will come. As part of runtime
suspend we will queue the stop rx sequencer work in GSI mode before even
initialization of workqueue is done and due to this we will get the NULL
pointer dereference and system will crash, hence to avoid this issue move
the initialization of workqueue in probe function.
- If we allocate the Tx and Rx channel before GENI init sequence done then
we will get the malformed TRE global error and hence to avoid this do
not start the Rx sequencer before the port setup is done for GSI mode.
- Destroy the workqueue in remove function and check the return condition
for allocation of Tx, RX and Event channel to avoid the NULL pointer
derefernce issues if channel allocation is failed.

Change-Id: I45db11ab7cd4da03fae9926f19118c54e363b409
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>