serial: msm_geni_serial: Return the error code instead of success
In rare cases we see that runtime_suspend and UART shutdown function
can call stop rx sequencer at the same time, so if one stop rx
sequencer is in progress then we will returning 0 from the stop rx
sequencer. In runtime_suspend function we will think rx sequencer is
stopped but in actual case it is not stopped yet and we will continue
the runtime_suspend operation and as part of RT suspend we will turn
off the resources and eventually it will lead to the NOC issues, so
to avoid this issue return the -EBUSY error instead of success.

Change-Id: I455dfddfde7ea4eceaf4f5ed1b44bec8bf32f878
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>