usb: gadget: u_serial: Prevent list corruption on queued_write_pool
In execution of gs_start_tx the usb_request are fetched from the
write_pool and perform gs_send_packet for the req_buf. If the length
isn't equal to 0 then it will go ahead and move it to queued_write_pool
and remove it from write_pool. After this, the request is queued to the
controller.
Now, consider if the next request's length is equal to 0 then the while
loop breaks and returns from gs_start_tx.
When the gs_start_tx starts again and iterates over the list. It finds
the list head now points to (next)->next request i.e skipping the
second request in the write_pool.

Fix this by moving the request to queue_write_pool as soon as it is
extracted from the write_pool. This way even if the while loop breaks
the request would be in the queued pool.

Change-Id: I945fd4d940e7ee03fb2ba3c46ac1e31e51030141
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>