drm/virtgpu: Fix vendor feature bit
Do not check vendor feature bit when CONFIG_MSM_VIRTIO_HAB
is not enabled, otherwise it will trigger BUG_ON because the
vendor feature bit is not in driver feature table in this case.

This change also removes the hack to force vendor bit because
it has been properly handled by host device code.

Change-Id: I5c57055e54b4112a7d965fb736b4eb9c17dd6311
Signed-off-by: Hai Li <quic_hali@quicinc.com>