oplus_chg: disable non-swarp strategy policies
This mechanism serves to limit the charging current (not input current)
depending on the surface temperature, but it fails in practice for so
many different reasons.

First, the input current is still handled separately, which is the major
producer of heat when the phone is in use anyways.

Second, the threshold is set too low for a hot chip like the sm8350, and
this will throttle the charging current to 500 mA hard even on light-
weight uses. While input current is handled differently, so the battery
won't actually get discharged, it'll charge very slowly while most of the
heat is still being produced from the input current anyways.

Third, this is updated very rarely, often in 6-10 minutes range. This
means, even after the device has cooled down, it'll take a long time for
the charging speed to recover.

Fourth, the actual dynamic current adjustment is still done separately
by monitoring the battery temperature. The only reason this exists is to
keep the surface temperature in check, but it instead causes huge user
inconveniences, often resulting in charging the battery at < 1A.

For better practical and balanced user experience, remove static strategy
policies.

Super Warp charging is excluded here as it is handled differently.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>